+++ packages/jest-config/src/getMaxWorkers.ts
@@ -11,6 +11,7 @@ import {Config} from '@jest/types';
 export default function getMaxWorkers(
   argv: Partial<Pick<Config.Argv, 'maxWorkers' | 'runInBand' | 'watch'>>,
 ): number {
+  const cpus = os.cpus() ? os.cpus().length : 1;
   if (argv.runInBand) {
     return 1;
   } else if (argv.maxWorkers) {
@@ -18,7 +19,6 @@ export default function getMaxWorkers(
     const maxWorkers = (argv.maxWorkers as unknown) as number | string;
     const parsed = parseInt(maxWorkers as string, 10);
 
-    const cpus = os.cpus().length;
     if (
       typeof maxWorkers === 'string' &&
       maxWorkers.trim().endsWith('%') &&
@@ -29,10 +29,9 @@ export default function getMaxWorkers(
       return workers >= 1 ? workers : 1;
     }
 
-    return Math.max(Math.min(cpus, parsed), 1);
+    return Math.max(Math.min(cpus - 1, parsed), 1);
   } else {
     // In watch mode, Jest should be unobtrusive and not use all available CPUs.
-    const cpus = os.cpus() ? os.cpus().length : 1;
     return Math.max(argv.watch ? Math.floor(cpus / 2) : cpus - 1, 1);
   }
 }